+++ lib/origin_recognizer.dart
@@ -4,21 +4,14 @@ import 'origin_triggers.dart';
 class OriginScaleRecognizer extends ScaleGestureRecognizer {
   OriginScaleRecognizer({super.supportedDevices, super.dragStartBehavior = DragStartBehavior.down});
 
-  Map<OriginStart, OriginBounds> gestures = {};
+  List<OriginGesture> gestures = [];
 
   final trackedPointers = <int>{};
   Offset _totalDelta = .zero;
 
-  bool get _hasSinglePointerGestures => gestures.keys.any((s) => switch (s) {
-    .left || .right || .up || .down => true,
-    _ => false,
-  });
+  bool get _hasSinglePointerGestures => gestures.any((g) => g.start.any((s) => s.isOne));
 
-  bool get _hasMultiPointerGestures => gestures.keys.any((s) => switch (s) {
-    .twoLeft || .twoRight || .twoUp || .twoDown ||
-    .pinchIn || .pinchOut => true,
-    _ => false,
-  });
+  bool get _hasMultiPointerGestures => gestures.any((g) => g.start.any((s) => s.isTwo || s.isScale));
 
   @override
   void addAllowedPointer(PointerDownEvent event) {
@@ -64,13 +57,15 @@ class OriginScaleRecognizer extends ScaleGestureRecognizer {
     super.resolve(disposition);
   }
 
+  bool _hasStart(OriginStart s) => gestures.any((g) => g.start.contains(s));
+
   bool _matchesSinglePointer() {
     final dx = _totalDelta.dx;
     final dy = _totalDelta.dy;
-    if (dx > 2 && gestures.containsKey(OriginStart.right)) return true;
-    if (dx < -2 && gestures.containsKey(OriginStart.left)) return true;
-    if (dy > 2 && gestures.containsKey(OriginStart.down)) return true;
-    if (dy < -2 && gestures.containsKey(OriginStart.up)) return true;
+    if (dx > 2 && _hasStart(.right)) return true;
+    if (dx < -2 && _hasStart(.left)) return true;
+    if (dy > 2 && _hasStart(.down)) return true;
+    if (dy < -2 && _hasStart(.up)) return true;
     return false;
   }